drum-transcription/requirements.txt
# Core dependencies
torch>=1.9.0
torchaudio>=0.9.0
numpy>=1.20.0
scipy>=1.7.0
matplotlib>=3.4.0
librosa>=0.8.1
pyyaml>=5.4.0
tqdm>=4.61.0
soundfile>=0.10.3
pandas>=1.3.0
mido>=1.2.10
pretty_midi>=0.2.9

# HTDemucs dependencies
demucs>=3.0.0

# Optional dependencies for notebooks
jupyter>=1.0.0
ipython>=7.26.0
seaborn>=0.11.2

# For testing
pytest>=6.2.5
pytest-cov>=2.12.1 